Refactor function to make it more general
The "filter out" and "group by" features have been removed from the
function, this way it only increases the available data.

They will be reimplemented in the future as a separated feature to
study dynamic coefficient of friction, alongside with other features
that wouldn't be possible with a data set that has been filtered and
grouped.